+++ src/Vortice.Direct3D12/ID3D12DeviceChild.cs
@@ -2,30 +2,28 @@
 // Distributed under the MIT license. See the LICENSE file in the project root for more information.
 
 using System;
-using System.Runtime.CompilerServices;
+using SharpGen.Runtime;
 
 namespace Vortice.Direct3D12
 {
     public partial class ID3D12DeviceChild
     {
-        private ID3D12Device _device;
-
-        public ID3D12Device Device
+        public Result GetDevice<T>(out T device) where T : ID3D12Device
         {
-            get
+            Result result = GetDevice(typeof(T).GUID, out IntPtr nativePtr);
+            if (result.Failure)
             {
-                if (_device == null)
-                {
-                    _device = GetDevice<ID3D12Device>();
-                }
-
-                return _device;
+                device = default;
+                return result;
             }
+
+            device = FromPointer<T>(nativePtr);
+            return result;
         }
 
         public T GetDevice<T>() where T : ID3D12Device
         {
-            if (GetDevice(typeof(T).GUID, out var nativePtr).Failure)
+            if (GetDevice(typeof(T).GUID, out IntPtr nativePtr).Failure)
             {
                 return default;
             }